brightness changing

+++ app/src/main/java/com/example/annyslamp/core/event/LampEvent.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.Color
 
 sealed class LampEvent {
     object Toggle : LampEvent()
-    data class SetBrightness(val value: Int) : LampEvent()
+    data class SetBrightness(val value: Float) : LampEvent()
     data class SetColor(val color: Color) : LampEvent()
     data class SetMode(val mode: LampMode) : LampEvent()
 }
